+++ example/lib/ui/profile_screen.dart
@@ -1,10 +1,11 @@
+import 'package:example/ui/home_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:router_management/router_management.dart';
 
 class ProfileScreen extends StatelessWidget {
   const ProfileScreen({Key? key}) : super(key: key);
 
-  static const path = '/profile';
+  static const path = '${HomeScreen.path}/profile';
 
   static const name = 'Profile';
 
@@ -12,9 +13,14 @@ class ProfileScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     final arguments = context.arguments;
 
-    return Scaffold(
-      body: Center(
-        child: Text('Hello, ${arguments.query['name']!}'),
+    return GestureDetector(
+      onTap: () {
+        context.navigator.pop();
+      },
+      child: Scaffold(
+        body: Center(
+          child: Text('Hello, ${arguments.query['name']}'),
+        ),
       ),
     );
   }