+++ lib/router_management.dart
@@ -1,11 +1,11 @@
 library router_management;
 
+export 'src/data/extensions/navigation_extension.dart';
+export 'src/data/extensions/page_arguments_extension.dart';
+export 'src/data/models/navigation_page.dart';
+export 'src/data/models/page_arguments.dart';
+export 'src/data/mixins/navigation.dart';
 export 'src/data/services/navigation_service.dart';
-export 'src/domain/models/navigation_page.dart';
-export 'src/domain/models/page_arguments.dart';
-export 'src/domain/services/navigation.dart';
-export 'src/domain/validators/page_validator.dart';
-export 'src/domain/widgets/page_widget.dart';
-export 'src/infrastructure/extensions/navigation_extension.dart';
-export 'src/infrastructure/extensions/page_arguments_extension.dart';
+export 'src/data/validators/page_validator.dart';
 export 'src/presentation/ui/navigation_router.dart';
+export 'src/presentation/widgets/page_widget.dart';