+++ lib/src/data/models/page_arguments.dart
@@ -29,5 +29,6 @@ class PageArguments {
   /// the current list of queries as a [Map<String,List<String>>] object
   late final queries = uri.queryParametersAll;
 
+  /// Is used to create a new instance of [PageArguments]
   PageArguments(this.uri, [this.data]);
 }