+++ lib/src/data/services/navigation_parser.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 
-import '../../domain/models/page_arguments.dart';
 import '../mixins/navigation_parser_mixin.dart';
+import '../models/page_arguments.dart';
 
 class NavigationParser extends RouteInformationParser<PageArguments>
     with NavigationParserMixin {