+++ lib/src/data/services/navigation_service.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/widgets.dart';
 
-import '../../domain/models/page_arguments.dart';
-import '../../domain/services/navigation.dart';
-import '../mixins/navigation_delegate_mixin.dart';
+import '../mixins/navigation.dart';
+import '../../presentation/mixins/navigation_delegate_mixin.dart';
 import '../mixins/navigation_mixin.dart';
+import '../models/page_arguments.dart';
 
 /// [NavigationService] is the core class that is used to get the actual
 /// [Navigation] instance
@@ -14,6 +14,8 @@ class NavigationService extends RouterDelegate<PageArguments>
         Navigation,
         NavigationMixin,
         NavigationDelegateMixin {
+
+  /// Is used internally
   NavigationService._();
 
   /// [NavigationService.instance] is the implementation of navigator 2.0