+++ lib/src/data/validators/page_validator.dart
@@ -1,10 +1,11 @@
+import '../mixins/navigation.dart';
 import '../models/navigation_page.dart';
 import '../models/page_arguments.dart';
-import '../services/navigation.dart';
 
 /// [PageValidator] can be used to validate any pages with
 /// [NavigationPage.validators]
 mixin PageValidator {
+  /// Is used to validate the page when it will be activated by a system call
   Future<bool> call(
     NavigationPage page,
     Navigation navigation,