+++ lib/src/presentation/pages/page_route_navigation.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 import 'page_settings.dart';
@@ -28,13 +29,12 @@ class PageRouteNavigation extends PageRoute<void> {
   @override
   Widget buildTransitions(BuildContext context, Animation<double> animation,
       Animation<double> secondaryAnimation, Widget child) {
-    if (settings.transitionsBuilder != null) {
-      return settings.transitionsBuilder!(
-        context,
-        animation,
-        secondaryAnimation,
-        child,
-      );
+    if(animation.isCompleted || secondaryAnimation.isCompleted) return child;
+    
+    final transitionsBuilder = settings.transitionsBuilder;
+
+    if (transitionsBuilder != null) {
+      return transitionsBuilder(context, animation, secondaryAnimation, child);
     }
 
     final theme = Theme.of(context).pageTransitionsTheme;