+++ lib/src/presentation/pages/page_settings.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/widgets.dart';
 
-import '../../domain/models/page_arguments.dart';
+import '../../data/models/page_arguments.dart';
 import 'page_route_navigation.dart';
 
 class PageSettings<T> extends Page<void> {
@@ -12,6 +12,7 @@ class PageSettings<T> extends Page<void> {
   final Duration transitionDuration;
   final RouteTransitionsBuilder? transitionsBuilder;
   final Completer<T?>? completer;
+  final bool isInitialPage;
 
   PageSettings(
     String path,
@@ -25,6 +26,7 @@ class PageSettings<T> extends Page<void> {
     this.transitionsBuilder,
     bool isCompleted,
   )   : completer = isCompleted ? null : Completer(),
+        isInitialPage = false,
         super(
           key: ValueKey<String>(path),
           name: name,
@@ -39,6 +41,7 @@ class PageSettings<T> extends Page<void> {
         transitionDuration = const Duration(milliseconds: 400),
         transitionsBuilder = null,
         completer = null,
+        isInitialPage = true,
         super(key: ValueKey<String>(path));
 
   @override
@@ -50,7 +53,5 @@ class PageSettings<T> extends Page<void> {
   String get path => key.value;
 
   @override
-  Route<void> createRoute(BuildContext context) {
-    return PageRouteNavigation(this);
-  }
+  Route<void> createRoute(BuildContext context) => PageRouteNavigation(this);
 }